requirements.txt
google-endpoints==2.3.1
google-endpoints-api-management==1.3.0
google-auth==1.1.1
